djangomission/requirements.txt
asgiref==3.2.10
Django==3.1.14
pytz==2022.2.1
sqlparse==0.4.2
djangorestframework==3.13.1
ipython
django-extensions
